Made some changes to dependent source files - change double to float

+++ main_project/component/currentControlledCurrentSource.hpp
@@ -8,8 +8,7 @@
 
 class CurrentControlledCurrentSource: public CurrentControlledSource{
 public:
-    CurrentControlledCurrentSource(string _name, vector<string> args, vector<double> extraInfo);
-    // CurrentControlledCurrentSource(string _name, double _voltage, int n1, int n2, int nc1, int nc2);
+    CurrentControlledCurrentSource(string _name, vector<string> args, vector<float> extraInfo);
     ~CurrentControlledCurrentSource() = default;
 };
