+++ main_project/component/voltageControlledCurrentSource.hpp
@@ -8,8 +8,7 @@
 
 class VoltageControlledCurrentSource: public VoltageControlledSource{
 public:
-    VoltageControlledCurrentSource(string _name, vector<string> args, vector<double> extraInfo);
-    // VoltageControlledVoltageSource(string _name, double _voltage, int n1, int n2, int nc1, int nc2);
+    VoltageControlledCurrentSource(string _name, vector<string> args, vector<float> extraInfo);
     ~VoltageControlledCurrentSource() = default;
 };
