+++ main_project/component/voltageControlledSource.cpp
@@ -3,7 +3,7 @@
 
 #include "voltageControlledSource.hpp"
 
-VoltageControlledSource::VoltageControlledSource(string name, vector<string> args, vector<double> extraInfo)
+VoltageControlledSource::VoltageControlledSource(string name, vector<string> args, vector<float> extraInfo)
     :Component{name}
 {
     int n1 = stoi(args[0]);
@@ -16,11 +16,6 @@ VoltageControlledSource::VoltageControlledSource(string name, vector<string> arg
     gain = stod(args[4]);
 }
 
-// VoltageControlledSource::VoltageControlledSource(string _name, double _voltage, int n1, int n2, int nc1, int nc2)
-//     :Component{_name}{
-//     setupBasic(n1, n2, nc1, nc2);
-// }
-
 VoltageControlledSource::~VoltageControlledSource(){}
 
 void VoltageControlledSource::setupBasic(int n1, int n2, int nc1, int nc2){
@@ -30,7 +25,7 @@ void VoltageControlledSource::setupBasic(int n1, int n2, int nc1, int nc2){
     nodes.push_back(nc2);
 }
 
-double VoltageControlledSource::getGain() const{
+float VoltageControlledSource::getGain() const{
     return gain;
 }
