+++ main_project/component/voltageControlledVoltageSource.cpp
@@ -3,14 +3,8 @@
 
 #include "voltageControlledVoltageSource.hpp"
 
-VoltageControlledVoltageSource::VoltageControlledVoltageSource(string name, vector<string> args, vector<double> extraInfo)
+VoltageControlledVoltageSource::VoltageControlledVoltageSource(string name, vector<string> args, vector<float> extraInfo)
     :VoltageControlledSource(name, args, extraInfo)
 {
     types.push_back(componentType::voltageSource);
 }
-
-// VoltageControlledVoltageSource::VoltageControlledVoltageSource(string _name, double _voltage, int n1, int n2, int nc1, int nc2)
-//     :VoltageControlledSource(_name, _voltage, n1, n2, nc1, nc2)
-// {
-//     types.push_back(componentType::voltageSource);
-// }
