+++ main_project/component/voltageControlledVoltageSource.hpp
@@ -8,8 +8,7 @@
 
 class VoltageControlledVoltageSource: public VoltageControlledSource{
 public:
-    VoltageControlledVoltageSource(string _name, vector<string> args, vector<double> extraInfo);
-    // VoltageControlledVoltageSource(string _name, double _voltage, int n1, int n2, int nc1, int nc2);
+    VoltageControlledVoltageSource(string _name, vector<string> args, vector<float> extraInfo);
     ~VoltageControlledVoltageSource() = default;
 };
